requirements.txt
Django~=4.2.7
numpy~=1.24.3
mysql-connector-python~=8.2.0
scipy~=1.10.1
pyproj~=3.5.0
requests~=2.27.1
djangorestframework~=3.14.0
drf_yasg~=1.21.7